#286 : remove BuildConfig references
