+++ common/compose/build.gradle.kts
@@ -24,10 +24,6 @@ plugins {
 
 android {
     namespace = "dev.bwaim.loteria.compose"
-
-    buildFeatures {
-        buildConfig = true
-    }
 }
 
 dependencies {
