+++ common/compose/src/debug/java/dev.bwaim.loteria.compose/Debug.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2022 Dev Bwaim team
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@file:Suppress("NOTHING_TO_INLINE")
+
+package dev.bwaim.loteria.compose
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.remember
+import timber.log.Timber
+
+@PublishedApi()
+internal class Ref(var value: Int)
+
+/**
+ * An effect which logs the number compositions at the invoked point of the slot table.
+ *
+ * This is an inline function to act as like a C-style #include to the host composable function.
+ * That way we track it's compositions, not this function's compositions.
+ *
+ * @param tag Log tag used for [Logger.d].
+ */
+@Composable
+public inline fun LogCompositions(tag: String) {
+    val ref = remember { Ref(0) }
+    SideEffect { ref.value++ }
+    Timber.d(tag, "Compositions: ${ref.value}")
+}
