+++ common/http/build.gradle.kts
@@ -33,10 +33,6 @@ android {
             )
         }
     }
-
-    buildFeatures {
-        buildConfig = true
-    }
 }
 
 dependencies {
