+++ common/http/src/debug/java/dev/bwaim/loteria/http/extension/OkHttpClientBuilderExtension.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Dev Bwaim team
+ * Copyright 2023 Dev Bwaim team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-plugins {
-    id("loteriamexicana.android.library")
-    id("loteriamexicana.android.library.jacoco")
-    id("loteriamexicana.hilt")
-    id("loteriamexicana.spotless")
-}
 
-android {
-    namespace = "dev.bwaim.loteria.imageloading"
-}
+package dev.bwaim.loteria.http.extension
 
-dependencies {
-    implementation(project(":common:coroutines:coroutines"))
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 
-    implementation(libs.androidx.core.library)
-
-    api(libs.coil.library)
+public fun OkHttpClient.Builder.addLogging(): OkHttpClient.Builder {
+    val loggingInterceptor = HttpLoggingInterceptor()
+        .setLevel(HttpLoggingInterceptor.Level.BODY)
+    return addInterceptor(loggingInterceptor)
 }
