+++ common/http/src/main/java/dev/bwaim/loteria/http/di/HttpModule.kt
@@ -22,12 +22,11 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import dev.bwaim.loteria.http.BuildConfig
+import dev.bwaim.loteria.http.extension.addLogging
 import dev.bwaim.loteria.http.interceptors.connectivity.ConnectivityInterceptor
 import dev.bwaim.loteria.http.interceptors.emptybody.EmptyBodyInterceptor
 import okhttp3.Cache
 import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import java.io.File
 import java.util.concurrent.TimeUnit
 import javax.inject.Singleton
@@ -49,13 +48,7 @@ internal object HttpModule {
         .writeTimeout(TIMEOUT, TimeUnit.MILLISECONDS)
         .addNetworkInterceptor(connectivityInterceptor)
         .addInterceptor(emptyBodyInterceptor)
-        .apply {
-            if (BuildConfig.DEBUG) {
-                val loggingInterceptor = HttpLoggingInterceptor()
-                    .setLevel(HttpLoggingInterceptor.Level.BODY)
-                addInterceptor(loggingInterceptor)
-            }
-        }
+        .addLogging()
         .cache(Cache(File(context.cacheDir, "http_cache"), 50L * 1024L * 1024L))
         .build()
 }
