+++ common/initializers/src/debug/AndroidManifest.xml
@@ -26,7 +26,7 @@
             tools:node="merge">
 
             <meta-data
-                android:name="dev.bwaim.loteria.initializers.TimberInitializer"
+                android:name="dev.bwaim.loteria.initializers.StrictModeInitializer"
                 android:value="androidx.startup" />
         </provider>
